scrapers/requirements.txt
bs4
requests
boto3
python-dotenv